requirements.txt
certifi==2020.6.20
chardet==3.0.4
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jws==0.1.3
oauth2client==4.1.2
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
Pyrebase4==4.3.0
python-jwt==2.0.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
urllib3==1.25.10
